+++ pkgs/http/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * `BrowserClient` throws an exception if `send` is called after `close`.
 * No longer depends on package:path.
+* If `no_default_http_client=true` is set in the environment then disk usage
+  is reduced in some circumstances.
 
 ## 0.13.5